% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densest.R
\name{ash_density}
\alias{ash_density}
\title{Density Estimate using Average Shifted Histogram Method}
\usage{
ash_density(x, v, h0 = 0.1, N = 9)
}
\arguments{
\item{x}{Numeric or Vector Value at which the density has to be estimated}

\item{v}{Vector of Numeric observations}

\item{h0}{Numeric; Bandwidth; default value is 0.1}

\item{N}{Integer; Number of averages taken; default value 9}
}
\value{
Returns the density estimate vector
}
\description{
Obtain densiy estimate of a vector using average shifted histograms for removing dependence on choice of origin
}
\author{
Mainak Manna, Samya Praharaj
}
