% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densest.R
\name{ash_density_int}
\alias{ash_density_int}
\title{Density Estimate using Integrated Shifted Histogram Method}
\usage{
ash_density_int(x, v, h = 0.1)
}
\arguments{
\item{x}{Numeric or Vector Value at which the density has to be estimated}

\item{v}{Vector of Numeric observations}
}
\value{
Returns the density estimate vector
}
\description{
Obtain density estimate of a vector using integrated shifted histograms for removing dependence on choice of origin
}
\author{
Mainak Manna, Samya Praharaj
}
