% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densest.R
\name{Asymp}
\alias{Asymp}
\title{Plotting asymptotic normality of kernel density estimates}
\usage{
Asymp(
  u = 0,
  n = c(10, 30, 70, 150),
  expr = NULL,
  rs = rnorm,
  ker_f = dnorm,
  h0 = NULL,
  ...
)
}
\arguments{
\item{u}{Numeric value; Distance from location}

\item{n}{Numeric value; Sample Size}

\item{expr}{expression for density; Default value is NULL}

\item{rs}{function from which random sample of the specific distribution will be generated.}

\item{ker_f}{kernel function}

\item{h0}{Numeric value; Bandwidth}
}
\value{
Returns the plot of kernel density estimates
}
\description{
This function can be used to observe rate of convergence to normality by varying sample  size, location , underlying distribution and kernel function.
}
\author{
Mainak Manna, Samya Praharaj
}
