% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densest.R
\name{beta_density}
\alias{beta_density}
\title{Density Estimate using Beta pdf}
\usage{
beta_density(a, vec, isplot = FALSE, addplot = FALSE, col_plot = 1)
}
\arguments{
\item{a}{Numeric or Vector Value at which the density has to be estimated}

\item{vec}{Vector of Numeric observations}

\item{isplot}{Logical; Indicating whether to plot the estimated density; default value FALSE}

\item{addplot}{Logical; If true add to an existing plot; default value FALSE}

\item{col_plot}{Colour of the Plot}
}
\value{
If isplot is FALSE then returns the density estimate vector; Otherwise plots the estimated density
}
\description{
Obtain densiTy estimate of a vector using beta pdf
}
\author{
Mainak Manna, Samya Praharaj
}
