% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densest.R
\name{cosine_kern}
\alias{cosine_kern}
\title{Cosine Kernel}
\usage{
cosine_kern(x)
}
\arguments{
\item{x}{Numeric or Vector Value at which the density has to be estimated}
}
\description{
Obtain value of cosine kernel in given vector
}
\author{
Snehashis Dey, Mainak Manna
}
