% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densest.R
\name{et.l1}
\alias{et.l1}
\title{Calculating MISE for L1 loss using Monte Carlo Simulation}
\usage{
et.l1(
  func1 = rnorm,
  func2 = dnorm,
  Func = beta_density,
  l1 = -Inf,
  u1 = Inf,
  ...
)
}
\arguments{
\item{func1}{function to simulate data from required distribution; Default value rnorm}

\item{func2}{density function of required distribution}

\item{Func}{density estimate}

\item{l1, u1}{lower and upper bound of support of the kernel function}
}
\value{
Returns the approximate MISE for L1 loss using Monte Carlo Simulation
}
\description{
To approximate MISE for L1 loss using Monte Carlo Simulation for different distributions
}
\author{
Mainak Manna, Samya Praharaj
}
