% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densest.R
\name{histogram_density}
\alias{histogram_density}
\title{Density Estimate using Histogram method}
\usage{
histogram_density(x, v, x0 = 0, h = 0.1)
}
\arguments{
\item{x}{Numeric or Vector Value at which the density has to be estimated}

\item{v}{Vector of Numeric observations}

\item{x0}{Numeric; Origin ; default vlue is 0}

\item{h}{Numeric; Bandwidth; default value is 0.1}
}
\value{
Returns the density estimate vector
}
\description{
Obtain densiy estimate of a vector using histograms
}
\author{
Snehashis De, Samya Praharaj
}
