% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densest.R
\name{kern_density}
\alias{kern_density}
\title{To Find density estimates for a given data}
\usage{
kern_density(x, v, h = 0.1, func = dnorm, ...)
}
\arguments{
\item{x}{Numeric or Vector Value at which the density has to be estimated}

\item{v}{Vector of Observations}

\item{h}{bandwidth}

\item{func}{kernel function to be used}
}
\description{
To Find density estimates for a given data for different kernel functions
}
\author{
Mainak Manna, Samya Praharaj
}
