% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densest.R
\name{M_0}
\alias{M_0}
\title{Integrated Square Error}
\usage{
M_0(v, l1 = -Inf, u1 = Inf, h = 0.1, kern = dnorm, ...)
}
\arguments{
\item{v}{Vector value; The vector of observations}

\item{l1, u1}{lower and upper bound of support of the kernel function with default values -Inf and Inf respectively}

\item{h}{Vector value; Bandwidth with default value 0.1}

\item{kern}{kernel function}
}
\value{
Returns the value of Integrated Square Error
}
\description{
This function can be used to find the integrated square error for different bandwidths
}
\author{
Mainak Manna, Samya Praharaj
}
