% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densest.R
\name{naive_kern}
\alias{naive_kern}
\title{Naive Kernel}
\usage{
naive_kern(x)
}
\arguments{
\item{x}{Numeric or Vector Value at which the density has to be estimated}
}
\value{
Returns the value of naive kernel in given vector
}
\description{
Obtain value of naive kernel in given vector
}
\author{
Mainak Manna, Samya Praharaj
}
