% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densest.R
\name{new_kern}
\alias{new_kern}
\title{Optimal Kernel Which can take Negative Values}
\usage{
new_kern(x)
}
\arguments{
\item{x}{Numeric or Vector Value at which the density has to be estimated}
}
\description{
Obtain value of a kernel which can take negative values in given vector
}
\author{
Mainak Manna, Samya Praharaj
}
