% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densest.R
\name{opt_kern}
\alias{opt_kern}
\title{Epanenchnikov Kernel for L2 loss}
\usage{
opt_kern(x)
}
\arguments{
\item{x}{Numeric or Vector Value at which the density has to be estimated}
}
\description{
Obtain value of Epanenchnikov kernel which is  Optimal kernel for L2 loss, in given vector
}
\author{
Snehashis Dey, Samya Praharaj
}
