% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densest.R
\name{shiftedexpo_kern}
\alias{shiftedexpo_kern}
\title{Shifted Exponential Kernel}
\usage{
shiftedexpo_kern(x)
}
\arguments{
\item{x}{Numeric or Vector Value at which the density has to be estimated}
}
\description{
Obtain value of shifted exponential kernel in given vector
}
\author{
Snehashis Dey, Mainak Manna
}
