% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densest.R
\name{triangular_kern}
\alias{triangular_kern}
\title{Triangular Kernel}
\usage{
triangular_kern(x)
}
\arguments{
\item{x}{Numeric or Vector Value at which the density has to be estimated}
}
\description{
Obtain value of triangular kernel in given vector
}
\author{
Snehashis Dey, Mainak Manna
}
